CODE CLEANUP and OPTIMIZATION
All important codes were cleaned up and organized. There was the addition of the autoBase which includes the implementation of an abstract class (although as of the time of upload it has yet to be tested). If functional, it should allow for more versatile code usage throughout the various programs. Additionally, I also used this time to flesh out the Typex Chart which will hopefully work as planned. File naming system has also been added to incorporate things that make more sense and are relevant to their programmed purpose.