Autonomous Expansion
Essentially, this update to the program adds an expansion to the autoBase class with encoder control for forward driving (without additional controllers). It also adds slight modifications to the individual autos and reintroduces the CONSTANTS class. There is also the additional program that uses mainly encoders however it will be tested tomorrow for both accuracy and precision.