ENCODER FIXES V.1
The program that I designed did not work as expected. I hae updated it and it works reliably for two two of the wheels. I attempted to make it work for the other two, but due to time constaints work has been paused until a later point in time. The code right now regarding the encoderTheory and ENCODER_FOUNDATION_BLUE remain a work in progress.